+++ src/Akeneo/Platform/Component/CatalogVolumeMonitoring/Volume/Service/VolumeAggregation.php
@@ -5,9 +5,8 @@ declare(strict_types=1);
 namespace Akeneo\Platform\Component\CatalogVolumeMonitoring\Volume\Service;
 
 use Akeneo\Platform\Component\CatalogVolumeMonitoring\Volume\Model\AggregatedVolume;
-use Akeneo\Platform\Component\CatalogVolumeMonitoring\Volume\Query\AverageMaxQuery;
-use Akeneo\Platform\Component\CatalogVolumeMonitoring\Volume\Query\CountQuery;
 use Akeneo\Platform\Component\CatalogVolumeMonitoring\Volume\Repository\AggregatedVolumeRepositoryInterface;
+use Psr\Log\LoggerInterface;
 
 /**
  * Aggregation of catalog volumes, for the queries too expensive to be executed live.
@@ -18,28 +17,12 @@ use Akeneo\Platform\Component\CatalogVolumeMonitoring\Volume\Repository\Aggregat
  */
 class VolumeAggregation
 {
-    /** @var AggregatedVolumeRepositoryInterface */
-    private $aggregatedVolumeRepository;
-
-    /** @var iterable */
-    private $countQueries;
-
-    /** @var iterable */
-    private $averageMaxQueries;
-
-    /**
-     * @param AggregatedVolumeRepositoryInterface $aggregatedVolumeRepository
-     * @param CountQuery[]                        $countQueries
-     * @param AverageMaxQuery[]                   $averageMaxQueries
-     */
     public function __construct(
-        AggregatedVolumeRepositoryInterface $aggregatedVolumeRepository,
-        iterable $countQueries,
-        iterable $averageMaxQueries
+        private AggregatedVolumeRepositoryInterface $aggregatedVolumeRepository,
+        private iterable $countQueries,
+        private iterable $averageMaxQueries,
+        private LoggerInterface $logger,
     ) {
-        $this->aggregatedVolumeRepository = $aggregatedVolumeRepository;
-        $this->countQueries = $countQueries;
-        $this->averageMaxQueries = $averageMaxQueries;
     }
 
     /**
@@ -47,6 +30,8 @@ class VolumeAggregation
      */
     public function aggregate(): void
     {
+        $this->logger->info('Volumes aggregation in progress.');
+
         foreach ($this->countQueries as $countQuery) {
             $countVolume = $countQuery->fetch();
             $aggregatedVolume = new AggregatedVolume(
@@ -62,14 +47,18 @@ class VolumeAggregation
             $averageMaxVolume = $averageMaxQuery->fetch();
             $aggregatedVolume = new AggregatedVolume(
                 $averageMaxVolume->getVolumeName(),
-                ['value' => [
-                    'max' => $averageMaxVolume->getMaxVolume(),
-                    'average' => $averageMaxVolume->getAverageVolume(),
-                ]],
+                [
+                    'value' => [
+                        'max' => $averageMaxVolume->getMaxVolume(),
+                        'average' => $averageMaxVolume->getAverageVolume(),
+                    ],
+                ],
                 new \DateTime('now', new \DateTimeZone('UTC'))
             );
 
             $this->aggregatedVolumeRepository->add($aggregatedVolume);
         }
+
+        $this->logger->info('Catalog volumes aggregation done.');
     }
 }